+++ CHANGELOG.md
@@ -9,15 +9,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- `YoutubeDownloader\Provider\Youtube\Provider::createFromOptions()` to create the Youtube Provider with an options array
-- `YoutubeDownloader\Provider\Youtube\VideoInfo::createFromStringWithOptions()` to create the Youtube VideoInfo with an options array
+- new `YoutubeDownloader\Config\TransformationConfig` for a BC friendly configuraion
+- new `YoutubeDownloader\Provider\Youtube\Provider::createFromOptions()` to create the Youtube Provider with an options array
+- new `YoutubeDownloader\Provider\Youtube\VideoInfo::createFromStringWithOptions()` to create the Youtube VideoInfo with an options array
 
 ### Changed
 
 - The mp3 downloader was improved and has no dependendy to aria2 anymore
+- Moved configuration `$config['ThumbnailImageMode']` to `$config['gui']['ThumbnailImageMode']`
+- Moved configuration `$config['VideoLinkMode']` to `$config['gui']['VideoLinkMode']`
+- Moved configuration `$config['showBrowserExtensions']` to `$config['gui']['showBrowserExtensions']`
 
 ### Deprecated
 
+- `YoutubeDownloader\Config` will be removed in 0.7, use `YoutubeDownloader\Config\TransformationConfig` instead
 - `YoutubeDownloader\Provider\Youtube\Provider::createFromConfigAndToolkit()` will be removed in 0.7, use `YoutubeDownloader\Provider\Youtube\Provider::createFromOptions()` instead
 - `YoutubeDownloader\Provider\Youtube\VideoInfo::createFromStringWithConfig()` will be removed in 0.7, use `YoutubeDownloader\Provider\Youtube\VideoInfo::createFromStringWithOptions()` instead
 - `YoutubeDownloader\Toolkit::validateVideoId()` isn't used anymore and will be removed in 0.7